mattpad_modular/mattpad/requirements.txt
customtkinter>=5.2.0
requests>=2.28.0
chardet>=5.2.0
pyspellchecker>=0.7.0
keyring>=24.0.0
openai>=1.0.0
anthropic>=0.18.0
Pillow>=10.0.0
pywin32>=306;sys_platform=="win32"
